+++ examples/example_enum.c
@@ -20,16 +20,17 @@ int main(int argc, char **argv) {
   luaA_enum_value(L, cards, CLUBS);
   luaA_enum_value(L, cards, SPADES);
   luaA_enum_value(L, cards, INVALID);
-  luaA_enum_value_name(L, cards, SPADES, "ESPADIA");
 
-  cards test = INVALID;
+  cards cval = SPADES;
+  const char* lval = "SPADES";
   
-  luaA_push(L, cards, &test);
+  luaA_push(L, cards, &cval);
+  printf("%i pushed as %s\n", cval, lua_tostring(L, -1));
   lua_pop(L, 1);
   
-  lua_pushstring(L, "INVALID");
-  luaA_to(L, cards, &test, -1);
-  printf("INVALID read back as %i\n", test); 
+  lua_pushstring(L, lval);
+  luaA_to(L, cards, &cval, -1);
+  printf("%s read back as %i\n", lval, cval); 
   lua_pop(L, 1);
   
   luaA_close(L);