+++ examples/example_unnested.c
@@ -6,20 +6,18 @@ int fib(int n) {
   return fib(n-1) + fib(n-2);
 }
 
+luaA_function_declare(fib, int, int);
+
 int main(int argc, char** argv) {
   
-  /* Init Lua & LuaAutoC */
   lua_State* L = luaL_newstate();
   luaA_open(L);
   
-  /* Register `fib` function */
-  luaA_function(L, fib, int, int);
+  luaA_function_register(L, fib, int, int);
   
-  /* Push integer onto stack and call `fib` */
   lua_pushinteger(L, 25);
   luaA_call(L, fib);
   
-  /* Print result & pop */
   printf("Result: %i\n", (int)lua_tointeger(L, -1));
   lua_pop(L, 1);